requirements.txt
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
sentry-sdk==1.3.1
psycopg2==2.9.1